moved search query to its own file, cleaning up code
